Extract shared runtime event bus contract
