+++ docs/tasks/20250924-todo-findings.md
@@ -19,3 +19,4 @@ Status: ✅ Completed 2025-09-24 — Added a canonical-case guard for device set
 • Low: Optional telemetry not surfaced yet in UI — dashboard reads zone.lighting/plantingPlan but backend snapshot lacks these aggregates; either add derived fields (backend) or guard UI (already guarded) and open an enhancement task. Files: src/backend/src/lib/uiSnapshot.ts, src/frontend/src/views/DashboardOverview.tsx, docs/system/ui-mplementation-spec.md
 Status: ✅ Completed 2025-09-25 — buildSimulationSnapshot now emits `zones[].lighting` coverage + plantingPlan snapshots, the dashboard metrics expose active plan capacity, and the UI spec documents the new payload contract.
 • Low: Path coupling in runtime imports — runtime/eventBus imports backend src modules directly; acceptable in monorepo but brittle for packaging; consider extracting a small shared package for EventBus contracts. Files: src/runtime/eventBus.ts, src/backend/src/lib/eventBus.ts, ADR 0002/0003
+Status: ✅ Completed 2025-09-25 — Extracted the event bus contract to `@runtime/eventBusCore`, updated runtime imports to use the shared module, re-exported it from the backend package, and documented the shared contract in ADR 0002/0003.
